Refactor + reworked resizing + conversion to doubles
TWEAKS:
-most values are of the double data type now, StorageValue and spinner were updated accordingly

-bar sorters can have thousands of bars now

-sorters dont reshuffle on window resize, added event to tell them when the window is resized

-created BarSorter to hold common code for both bar sorters

-created BarVisualizer for common code between bar visualizers

-bar sorters now have an adjustable AMOUNT of bars

-inverted image toggle button to have correct color/text

-implemented proper shuffling lmaooo you was doing it wack

-everything now uses a main JColorChooser in Customization GUI so it can store recents

+++ src/main/java/main/vcs/VisualComponent.java
@@ -3,18 +3,18 @@ package main.vcs;
 public class VisualComponent
 {
 	//the value that is compared to other VCs
-	private float value;
+	private double value;
 	
 	/**
 	 * these are what the sorters hold in their arrays, the things being sorted
 	 * VC for short
 	 */
-	public VisualComponent(float value)
+	public VisualComponent(double value)
 	{
 		this.value = value;
 	}
 	
-	public final float getValue()
+	public final double getValue()
 	{
 		return value;
 	}
